requirements.txt
openai
anthropic
google-genai
groq
python-dotenv
textblob
